refactor: split AudioActionDialog (797 lines) into 5 focused modules
Extract useAudioActionState hook, AudioActionAnalyzing, AudioActionInput,
and AudioActionPreview subcomponents. Main dialog now 160 lines.
